requirements.txt
frappe
PyPDF2
python-docx
requests
